% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My Bin Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{what it iterates through}

\item{n}{number}

\item{p}{probability}
}
\value{
barplot
}
\description{
My Bin Function
}
\examples{
mybin(iter=100,n=10, p=0.7)
}
