% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{val of n}

\item{iter}{what it iterates through}

\item{lambda}{value of lambda}

\item{...}{other parameters}
}
\value{
a graph
}
\description{
mycltp
}
\examples{
mycltp(n=2,iter=10000, lambda = 10)
}
