% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{funcoltion to be used}

\item{theta}{values}
}
\value{
plot of theta v. exp(z)
}
\description{
mymaxlikg
}
\examples{
mymaxlikg(theta=seq(0,1,length=10000))
}
