% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{value}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
list
}
\description{
myncurve
}
\examples{
myncurve(5, 5, 2)
}
