% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynTickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on plane}

\item{gamma}{probability of overbooking}

\item{p}{probability of selling out}
}
\value{
number of seats sold that could overflow
}
\description{
ntickets
}
\examples{
ntickets(N = 200, gamma = .02, p = .95)
}
